Refactor VueI18n creation to own module
This refactors the creation of the VueI18n
instance by encapsulating in an own ES module.
So the i18n instance can be imported to other pure
ES modules and will allow internationalization.
